+++ test/integration/datastore/async_methods/findAll.test.js
@@ -101,4 +101,15 @@ describe('DS.findAll(resourceName, params[, options]): ', function () {
 		assert.equal(lifecycle.serialize.callCount, 0, 'serialize should have been called');
 		assert.equal(lifecycle.deserialize.callCount, 1, 'deserialize should have been called');
 	});
+	it('should correctly propagate errors', function () {
+		$httpBackend.expectGET(/http:\/\/test\.angular-cache\.com\/posts\??/).respond(404, 'Not Found');
+
+		DS.findAll('post', {}).then(function () {
+			fail('Should not have succeeded!');
+		}, function (err) {
+			assert.equal(err.data, 'Not Found');
+		});
+
+		$httpBackend.flush();
+	});
 });