+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+##### 0.9.1 - xx May 2014
+
+###### Backwards compatible bug fixes
+- #69 - Failed requests should throw an error, not just return it
+
 ##### 0.9.0 - 22 May 2014
 
 ###### Breaking API changes